HW_04
Create a class named State that holds the following fields: a String
for the name of the state, an integer for the population, and two City
objects that hold data about the capital city and the most populous
city. The State constructor requires six parameters that represent the
names and populations of the state, its capital, and its most populous
city. Provide get methods for each field. Create the City class to be a
nonstatic, private inner class within the State class; the City class
contains a city’s name and population. Create a class to assign values
to and display values from two State objects. Save the files as
State.java and TestState.java.